refactor: replace workflow route with single page
