--- app/workflow/page.tsx
@@ -1,4 +1,0 @@
-export default function WorkflowPage() {
-  // Empty - the parallel routes handle content
-  return null;
-}
